requirements.txt
pip~=24.0
discord.py~=2.4.0
python-dotenv